pyproject.toml
[project]
name = "doc2image"
version = "0.1.0"
description = "Turn your documents into stunning AI-generated images."
authors = [
    {name = "Dylan Tintenfich",email = "tintenfichdylan05@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai (>=1.84.0,<2.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "streamlit (>=1.45.1,<2.0.0)",
    "pypdf (>=5.6.0,<6.0.0)",
    "ollama (>=0.5.1,<0.6.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "streamlit-aggrid (>=1.1.4.post1,<2.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
]
